requirements.txt
opencv-python==3.4.5.20
numpy>=1.16.3
scikit-learn>=0.21.1
Django==2.2
mysql-connector-python>=8.0.17